Поменял безымянные цифры в SearchTabPagerAdapter на константы Разделил некоторые классы в файлах Создал ImageProvider Добавил стратегии мокси Заменил тэги с простых строк на названия классов Добавил в .gitignore игнор дневника
