+++ .gitignore
@@ -58,4 +58,6 @@ fastlane/screenshots
 fastlane/test_output
 fastlane/readme.md
 
+# My files
 downloadables/
+diary.txt
