+++ app/build.gradle
@@ -44,4 +44,5 @@ dependencies {
     kapt 'com.github.bumptech.glide:compiler:4.8.0'
     implementation "io.reactivex.rxjava2:rxjava:2.2.1"
     implementation 'de.hdodenhof:circleimageview:3.0.0'
+    implementation "org.jetbrains.kotlin:kotlin-reflect:1.2.71"
 }
