+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/mainscreen/MainMvpViewInterface.kt
@@ -4,9 +4,9 @@ import com.arellomobile.mvp.MvpView
 import com.arellomobile.mvp.viewstate.strategy.AddToEndSingleStrategy
 import com.arellomobile.mvp.viewstate.strategy.StateStrategyType
 
-@StateStrategyType(AddToEndSingleStrategy :: class)
 interface MainMvpViewInterface : MvpView {
 
+    @StateStrategyType(AddToEndSingleStrategy :: class)
     fun selectScreen(id: Int)
 
 }
